Extract import routes
